Try to make citations dynamic with source info
